+++ test/unittests/test_id_to_iri.py
@@ -1,6 +1,7 @@
 """Unit tests for id to iri mapping"""
 
 import unittest
+import os
 
 from lxml import etree
 
@@ -9,6 +10,10 @@ from knora.dsplib.utils.id_to_iri import id_to_iri
 
 class TestIdToIri(unittest.TestCase):
     out_file = 'testdata/tmp/_test-id2iri-replaced.xml'
+
+    def setUp(self) -> None:
+        """Is executed before each test"""
+        os.makedirs('testdata/tmp', exist_ok=True)
 
     def test_invalid_xml_file_name(self):
         with self.assertRaises(SystemExit) as cm: